🚚 move rrweb-snapshot to the rrweb folder
