--- packages/rum-recorder/src/domain/rrweb-snapshot/index.ts
@@ -1,5 +1,0 @@
-import { serializeNodeWithId, transformAttribute, serializeDocument } from './snapshot'
-export * from './types'
-export * from './serializationUtils'
-
-export { serializeDocument, serializeNodeWithId, transformAttribute }
