+++ packages/rum-recorder/src/domain/rrweb/observer.ts
@@ -1,6 +1,6 @@
 import { monitor, callMonitored, throttle, DOM_EVENT, addEventListeners, addEventListener } from '@datadog/browser-core'
-import { getSerializedNodeId, hasSerializedNode } from '../rrweb-snapshot'
 import { nodeOrAncestorsShouldBeHidden, nodeOrAncestorsShouldHaveInputIgnored } from '../privacy'
+import { getSerializedNodeId, hasSerializedNode } from './serializationUtils'
 import {
   FocusCallback,
   HookResetter,
