+++ packages/rum-recorder/src/domain/rrweb/serializationUtils.spec.ts
