+++ packages/rum-recorder/src/domain/rrweb/serializationUtils.ts
