+++ packages/rum-recorder/src/domain/rrweb/serialize.spec.ts
@@ -1,7 +1,7 @@
 /* eslint-disable max-len */
 import { isIE } from '../../../../core/test/specHelper'
 import { IGNORED_NODE_ID } from './serializationUtils'
-import { absoluteToStylesheet, serializeNodeWithId } from './snapshot'
+import { absoluteToStylesheet, serializeNodeWithId } from './serialize'
 
 describe('absolute url to stylesheet', () => {
   const href = 'http://localhost/css/style.css'
