+++ packages/rum-recorder/src/domain/rrweb/serialize.ts
