+++ test/e2e/scenario/recorder.scenario.ts
@@ -1,7 +1,6 @@
 import { CreationReason, IncrementalSource, Segment } from '@datadog/browser-rum-recorder/cjs/types'
-import { InputData, StyleSheetRuleData } from '@datadog/browser-rum-recorder/cjs/domain/rrweb/types'
+import { InputData, StyleSheetRuleData, NodeType } from '@datadog/browser-rum-recorder/cjs/domain/rrweb/types'
 
-import { NodeType } from '@datadog/browser-rum-recorder/cjs/domain/rrweb-snapshot'
 import { createTest, bundleSetup, html, EventRegistry } from '../lib/framework'
 import { browserExecute } from '../lib/helpers/browser'
 import { flushEvents } from '../lib/helpers/sdk'
